+++ src/device/wocurtain.ts
@@ -3,7 +3,7 @@
  * wocurtain.ts: Switchbot BLE API registration.
  */
 import { SwitchbotDevice } from '../device.js';
-import { SwitchBotBLEModelName } from '../types.js';
+import { SwitchBotBLEModelFriendlyName, SwitchBotBLEModelName } from '../types.js';
 
 export class WoCurtain extends SwitchbotDevice {
   static parseServiceData(buf: Buffer, onlog: ((message: string) => void) | undefined) {
@@ -27,16 +27,13 @@ export class WoCurtain extends SwitchbotDevice {
     const lightLevel = (byte4 >> 4) & 0b00001111; // light sensor level (1-10)
     const deviceChain = byte4 & 0b00000111;
     const model = buf.subarray(0, 1).toString('utf8');
-    let modelName;
-    if (model === 'c') {
-      modelName = SwitchBotBLEModelName.Curtain;
-    } else {
-      modelName = SwitchBotBLEModelName.Curtain3;
-    }
+    const modelName = model === 'c' ? SwitchBotBLEModelName.Curtain : SwitchBotBLEModelName.Curtain3;
+    const modelFriendlyName = model === 'c' ? SwitchBotBLEModelFriendlyName.Curtain : SwitchBotBLEModelFriendlyName.Curtain3;
 
     const data = {
       model: model,
       modelName: modelName,
+      modelFriendlyName: modelFriendlyName,
       calibration: calibration,
       battery: battery,
       inMotion: inMotion,