+++ src/device/woiosensorth.ts
@@ -3,7 +3,7 @@
  * woiosensorth.ts: Switchbot BLE API registration.
  */
 import { SwitchbotDevice } from '../device.js';
-import { SwitchBotBLEModel, SwitchBotBLEModelName } from '../types.js';
+import { SwitchBotBLEModel, SwitchBotBLEModelFriendlyName, SwitchBotBLEModelName } from '../types.js';
 
 export class WoIOSensorTH extends SwitchbotDevice {
   static parseServiceData(serviceDataBuf: Buffer, manufacturerDataBuf: Buffer, onlog: ((message: string) => void) | undefined) {
@@ -36,6 +36,7 @@ export class WoIOSensorTH extends SwitchbotDevice {
     const data = {
       model: SwitchBotBLEModel.OutdoorMeter,
       modelName: SwitchBotBLEModelName.OutdoorMeter,
+      modelFriendlyName: SwitchBotBLEModelFriendlyName.OutdoorMeter,
       temperature: {
         c: temp_c,
         f: temp_f,