+++ src/device/wosensorth.ts
@@ -3,7 +3,7 @@
  * wosensorth.ts: Switchbot BLE API registration.
  */
 import { SwitchbotDevice } from '../device.js';
-import { SwitchBotBLEModel, SwitchBotBLEModelName } from '../types.js';
+import { SwitchBotBLEModel, SwitchBotBLEModelFriendlyName, SwitchBotBLEModelName } from '../types.js';
 
 export class WoSensorTH extends SwitchbotDevice {
   static parseServiceData(buf: Buffer, onlog: ((message: string) => void) | undefined) {
@@ -27,6 +27,7 @@ export class WoSensorTH extends SwitchbotDevice {
     const data = {
       model: SwitchBotBLEModel.Meter,
       modelName: SwitchBotBLEModelName.Meter,
+      modelFriendlyName: 'Meter',
       temperature: {
         c: temp_c,
         f: temp_f,
@@ -58,8 +59,9 @@ export class WoSensorTH extends SwitchbotDevice {
     const temp_f = Math.round(((temp_c * 9 / 5) + 32) * 10) / 10;
 
     const data = {
-      model: 'i',
-      modelName: 'WoSensorTHPlus',
+      model: SwitchBotBLEModel.MeterPlus,
+      modelName: SwitchBotBLEModelName.Meter,
+      modelfriendlyName: SwitchBotBLEModelFriendlyName.Meter,
       temperature: {
         c: temp_c,
         f: temp_f,